pyproject.toml
[tool.poetry]
name = "p4_classification_rh"
version = "0.1.0"
description = "Projet de classification RH utilisant le Machine Learning"
authors = ["MendyData <mendy.data@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.11"

# --- Data Science & ML ---
pandas = "^2.2.0"
numpy = "^1.26.0"
scikit-learn = "^1.4.0"
scipy = "^1.12.0"

# --- Visualisation ---
matplotlib = "^3.8.0"
seaborn = "^0.13.0"
plotly = "^5.18.0"

# --- Jupyter ---
jupyter = "^1.0.0"
jupyterlab = "^4.0.0"
ipykernel = "^6.28.0"

# --- Utilitaires ---
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"
tqdm = "^4.66.0"
loguru = "^0.7.2"

# --- Validation & Qualité des données ---
pydantic = "^2.5.0"
pandera = "^0.18.0"

[tool.poetry.group.dev.dependencies]
# --- Tests ---
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"

# --- Qualité du code ---
black = "^24.1.0"
isort = "^5.13.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"
pre-commit = "^3.6.0"

# --- Documentation ---
mkdocs = "^1.5.0"
mkdocs-material = "^9.5.0"

[tool.poetry.group.ml.dependencies]
# --- ML Avancé (optionnel) ---
xgboost = "^2.0.0"
lightgbm = "^4.3.0"
optuna = "^3.5.0"
mlflow = "^2.10.0"
shap = "^0.44.0"
imbalanced-learn = "^0.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# =============================================================================
# CONFIGURATION DES OUTILS
# =============================================================================

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | __pycache__
    | data
    | models
    | logs
)/
'''

[tool.isort]
profile = "black"
line_length = 88
skip = [".git", ".venv", "__pycache__", "data", "models", "logs"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
